统一大小写，规范部分类名 为Purchase 和 Use 添加了 Context 参数，可以方便进行某些处理 优化前端界面，补全了部分语言文件 优化商店页面的显示效果 暴露了UseHelper和PurchaseHelper，商品提供者可以在不进入商店页面的情况下进行购买和使用

+++ src/Api/Controller/history/RemoveHistoryController.php
@@ -3,9 +3,11 @@
 namespace Xypp\Store\Api\Controller\history;
 
 use Flarum\Api\Controller\AbstractDeleteController;
+use Flarum\Foundation\ValidationException;
 use Flarum\Http\RequestUtil;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Illuminate\Support\Arr;
+use Xypp\Store\Helper\StoreHelper;
 use Xypp\Store\PurchaseHistory;
 
 
@@ -20,7 +22,9 @@ class RemoveHistoryController extends AbstractDeleteController
         $history = PurchaseHistory::findOrFail($id);
         if ($actor->id != $history->user_id)
             $actor->assertCan("removePurchaseHistory");
-
-        $history->delete();
+        if (StoreHelper::applyExpire($history))
+            $history->delete();
+        else
+            throw new ValidationException(["error" => "xypp-store.forum.use_result.fail_expire"]);
     }
 }
